Add documentation to webots.cloud (#5734)
* nao md

* nao

* img

* rename

* move folder

* test

* test

* url

* change documentation link

* links

* file location

* add nao license

* insertbefore correct

* appearance

* change

* correct

* remove print

* unified font

* hide useless tabs

* cleanup

* blank

* typo

* cleanup

* change nao url

* additional links

* thumbnail

* grass thumbnail

* remove thumbnail

* omit docs images from test sources

* remove docs images from assets cache

* restore doc

* missing export

* restore reference

* cleanup

* test url

* doc

* yaml

* yaml

* missing export

* update md and for
