+++ projects/appearances/webots.yaml
@@ -1,0 +1,1 @@
+publish: true
