+++ projects/robots/pal_robotics/tiago_base/webots.yaml
@@ -1,0 +1,1 @@
+publish: true
