+++ projects/robots/pal_robotics/tiago_iron/webots.yaml
@@ -1,0 +1,1 @@
+publish: true
