+++ resources/web/wwi/viewer.js
@@ -1,12 +1,12 @@
 /* eslint no-extend-native: ["error", { "exceptions": ["String"] }] */
 /* global setup */
 /* global showdown */
-/* global hljs */
 
 'use strict';
 
 import {getGETQueryValue, getGETQueriesMatchingRegularExpression} from './request_methods.js';
 import {webots} from './webots.js';
+import {setupModalWindow, renderGraphs, highlightCode, updateModalEvents} from './proto_viewer.js';
 
 import WbImageTexture from './nodes/WbImageTexture.js';
 import WbPbrAppearance from './nodes/WbPbrAppearance.js';
@@ -313,112 +313,6 @@ function redirectImages(node) {
   }
 }
 
-function setupModalWindow() {
-  const doc = document.querySelector('#webots-doc');
-
-  // Create the following HTML tags:
-  // <div id="modal-window" class="modal-window">
-  //   <span class="modal-window-close-button">&times;</span>
-  //   <img class="modal-window-image-content" />
-  //   <div class="modal-window-caption"></div>
-  // </div>
-
-  const close = document.createElement('span');
-  close.classList.add('modal-window-close-button');
-  close.innerHTML = '&times;';
-  close.onclick = function() {
-    modal.style.display = 'none';
-  };
-
-  const loadImage = document.createElement('img');
-  loadImage.classList.add('modal-window-load-image');
-  loadImage.setAttribute('src', 'https://raw.githubusercontent.com/cyberbotics/webots/R2023a/resources/web/wwi/images/loading/load_animation.gif');
-
-  const image = document.createElement('img');
-  image.classList.add('modal-window-image-content');
-
-  const caption = document.createElement('div');
-  caption.classList.add('modal-window-caption');
-
-  const modal = document.createElement('div');
-  modal.setAttribute('id', 'modal-window');
-  modal.classList.add('modal-window');
-
-  modal.appendChild(close);
-  modal.appendChild(loadImage);
-  modal.appendChild(image);
-  modal.appendChild(caption);
-  doc.appendChild(modal);
-
-  window.onclick = function(event) {
-    if (event.target === modal) {
-      modal.style.display = 'none';
-      loadImage.style.display = 'block';
-      image.style.display = 'none';
-    }
-  };
-}
-
-function updateModalEvents(view) {
-  const modal = document.querySelector('#modal-window');
-  const image = modal.querySelector('.modal-window-image-content');
-  const loadImage = modal.querySelector('.modal-window-load-image');
-  const caption = modal.querySelector('.modal-window-caption');
-
-  // Add the modal events on each image.
-  const imgs = view.querySelectorAll('img');
-  for (let i = 0; i < imgs.length; i++) {
-    imgs[i].onclick = function(event) {
-      const img = event.target;
-      // The modal window is only enabled on big enough images and on thumbnail.
-      if (img.src.indexOf('thumbnail') === -1 && !(img.naturalWidth > 128 && img.naturalHeight > 128))
-        return;
-
-      // Show the modal window and the caption.
-      modal.style.display = 'block';
-      caption.innerHTML = (typeof this.parentNode.childNodes[1] !== 'undefined') ? this.parentNode.childNodes[1].innerHTML : '';
-
-      if (img.src.indexOf('.thumbnail.') === -1) {
-        // this is not a thumbnail => show the image directly.
-        image.src = img.src;
-        loadImage.style.display = 'none';
-        image.style.display = 'block';
-      } else {
-        // this is a thumbnail => load the actual image.
-        let url = img.src.replace('.thumbnail.', '.');
-        if (image.src === url) {
-          // The image has already been loaded.
-          loadImage.style.display = 'none';
-          image.style.display = 'block';
-          return;
-        } else {
-          // The image has to be loaded: show the loading image.
-          loadImage.style.display = 'block';
-          image.style.display = 'none';
-        }
-        // In case of thumbnail, search for the original png or jpg
-        image.onload = function() {
-          // The original image has been loaded successfully => show it.
-          loadImage.style.display = 'none';
-          image.style.display = 'block';
-        };
-        image.onerror = function() {
-          // The original image has not been loaded successfully => try to change the extension and reload it.
-          image.onerror = function() {
-            // The original image has not been loaded successfully => abort.
-            modal.style.display = 'none';
-            loadImage.style.display = 'block';
-            image.style.display = 'none';
-          };
-          url = img.src.replace('.thumbnail.jpg', '.png');
-          image.src = url;
-        };
-        image.src = url;
-      }
-    };
-  }
-}
-
 function applyAnchor() {
   const firstAnchor = document.querySelector("[name='" + localSetup.anchor + "']");
   if (firstAnchor) {
@@ -693,21 +587,6 @@ window.onpopstate = function(event) {
   setupUrl(document.location.href);
   getMDFile();
 };
-
-function highlightCode(view) {
-  const supportedLanguages = ['c', 'cpp', 'java', 'python', 'matlab', 'sh', 'ini', 'tex', 'makefile', 'lua', 'xml',
-    'javascript'];
-
-  for (let i = 0; i < supportedLanguages.length; i++) {
-    const language = supportedLanguages[i];
-    hljs.configure({languages: [ language ]});
-    const codes = document.querySelectorAll('.' + language);
-    for (let j = 0; j < codes.length; j++) {
-      const code = codes[j];
-      hljs.highlightBlock(code);
-    }
-  }
-}
 
 function resetRobotComponent(robot) {
   unhighlightX3DElement();
@@ -1199,19 +1078,6 @@ function applyTabs() {
   }
 }
 
-function renderGraphs() {
-  for (let id in window.mermaidGraphs) {
-    window.mermaidAPI.render(id, window.mermaidGraphs[id], function(svgCode, bindFunctions) {
-      document.querySelector('#' + id + 'Div').innerHTML = svgCode;
-      // set min-width to be 2/3 of the max-width otherwise the text might become too small
-      const element = document.querySelector('#' + id);
-      const style = element.getAttribute('style');
-      element.setAttribute('style',
-        style + ' min-width:' + Math.floor(0.66 * parseInt(style.split('max-width:')[1].split('px'))) + 'px;');
-    });
-  }
-}
-
 function applyAnchorIcons(view) {
   let elements = [];
   const tags = ['figcaption', 'h1', 'h2', 'h3', 'h4', 'h5'];
@@ -1682,7 +1548,7 @@ if (localSetup.book === 'blog') {
 }
 
 addContributionBanner();
-setupModalWindow();
+setupModalWindow('#webots-doc');
 applyToTitleDiv();
 getMDFile();
 getMenuFile();
