+++ scripts/web_component_studio/components.json
@@ -322,16 +322,6 @@
       }
     },
     {
-      "name": "nao",
-      "proto": "Nao",
-      "translation": "0 0 0.334",
-      "rotation": "0 0 1 2.23",
-      "viewpoint": {
-        "orientation": "-0.05023293193727536 -0.14589312777116745 0.9880242141861356 5.676132352811541",
-        "position": "-1.530061839136907 0.974027162193945 0.44353660021034313"
-      }
-    },
-    {
       "name": "ned",
       "proto": "Ned",
       "translation": "0 0 0",
